Renamed 'core' project to 'original'
