+++ src/original/contracts/MarbleDutchAuctionInterface.sol
