+++ src/original/contracts/MarbleNFTCandidateInterface.sol
