+++ src/original/contracts/MarbleNFTInterface.sol
