+++ src/original/deployments/008_set_minimal_price.js
