+++ src/original/externalABI/betaAuction.abi.1.0.js
