+++ src/original/externalABI/betaMarbleNFT.abi.1.0.js
